Move controls into subsystem
